Wui/api updates for experimental

+++ .internal/templates/scripts/build_installer.js
@@ -131,6 +131,9 @@ if [[ ! -f ./menu.sh ]]; then
   exit 2
 fi
 
+# Load meta data from installation
+source ./.internal/meta.sh
+
 #### Prebuild service scripts
 ${renderPrebuildScripts(options?.prebuildScripts)}
 
@@ -145,14 +148,22 @@ ${renderPostbuildScripts(options?.postbuildScripts)}
 
 #### End postbuild service scripts
 
-cp docker-compose-base.yml docker-compose.yml
+if [[ -f ./compose-override.yml ]]; then
+  echo "Merging 'compose-override.yml' with 'docker-compose-base.yml':"
+  echo "docker container run -it -v $(pwd)/compose-override.yml:/usr/iotstack_pycli/compose-override.yml:ro -v $(pwd)/docker-compose-base.yml:/usr/iotstack_pycli/docker-compose-base.yml:ro -v $(pwd)/docker-compose.yml:/usr/iotstack_pycli/docker-compose.yml -e \\"PYCLI_OVERRIDE_YML=compose-override.yml\\" -e \\"PYCLI_BASE_YML=docker-compose-base.yml\\" -e \\"PYCLI_OUTPUT_YML=docker-compose.yml\\" iostack_pycli:$VERSION /usr/local/bin/python3 /usr/iotstack_pycli/compose_override_entry.py"
+  docker container run -it -v $(pwd)/compose-override.yml:/usr/iotstack_pycli/compose-override.yml:ro -v $(pwd)/docker-compose-base.yml:/usr/iotstack_pycli/docker-compose-base.yml:ro -v $(pwd)/docker-compose.yml:/usr/iotstack_pycli/docker-compose.yml -e "PYCLI_OVERRIDE_YML=compose-override.yml" -e "PYCLI_BASE_YML=docker-compose-base.yml" -e "PYCLI_OUTPUT_YML=docker-compose.yml" iostack_pycli:$VERSION /usr/local/bin/python3 /usr/iotstack_pycli/compose_override_entry.py
+  echo ""
+  echo "Merge container exited. You can rerun merge in the future with:"
+  echo "  ./menu.sh --remerge-yaml-override --no-check"
+else
+  [ -f docker-compose.yml ] && rm docker-compose.yml
+  cp docker-compose-base.yml docker-compose.yml
+fi
 
 if [[ -f ./postbuild.sh ]]; then
   echo "Running postbuild script (${(options?.selectedServices ?? []).join(', ')}):"
   ./postbuild.sh ${(options?.selectedServices ?? []).join(' ')}
 fi
-
-# cat setup.json |  docker run -i  ubuntu /bin/bash -c 'cat' # for compose-override
 
 echo ""
 echo "Setup complete. You can start the stack with: "
